backend/requirements.txt
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
backports.zoneinfo==0.2.1
decorator==5.1.1
Django==3.2.9
django-cors-headers==4.0.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
executing==1.2.0
gunicorn==20.1.0
importlib-metadata==6.6.0
ipython==8.12.2
jedi==0.18.2
Markdown==3.4.3
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
PyJWT==1.7.1
psycopg2-binary==2.8.5
psycopg2==2.8.5
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
traitlets==5.9.0
typing-extensions==4.5.0
wcwidth==0.2.6
zipp==3.15.0
